requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
asyncssh
python-dotenv
pydantic-settings
httpx
websockets
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
cryptography
python-multipart
email-validator
alembic
openai
anthropic
